corrected my server action and my map function to display the cards
