+++ src/app/actions/index.ts
@@ -1,0 +1,1 @@
+export * from "./get-post.server"
